# doc/swagger.yaml
openapi: 3.1.0
info:
  title: Readmaster.ai API
  description: API for Readmaster.ai, a web-based reading assessment platform.
  version: 0.1.0
paths:
...
  /api/v1/parent/my-children:
    get:
      tags:
        - Parent - Child Monitoring
      summary: Parent List My Children
      description: Lists all children linked to the authenticated parent.
      operationId: parent_list_my_children_api_v1_parent_my_children_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponseDTO'
                type: array
                title: Response Parent List My Children Api V1 Parent My Children Get
      security:
        - OAuth2PasswordBearer: []
...
  /api/v1/parent/children:
    post:
      tags:
        - Parent - Account Management
      summary: Parent Create Child Account
      description: Allows an authenticated parent to create a new student account linked to them.
      operationId: parent_create_child_account_api_v1_parent_children_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentChildCreateRequest'
      responses:
        '201':
          description: Child account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized. Invalid or missing authentication token.
        '403':
          description: Forbidden. Authenticated user is not a parent.
        '422':
          description: Validation Error. Invalid input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parent/children/{child_id}/assignments:
    post:
      tags:
        - Parent - Assignments
      summary: Parent Assign Reading to Child
      description: Allows an authenticated parent to assign a specific reading material to one of their linked children.
      operationId: parent_assign_reading_to_child_api_v1_parent_children__child_id__assignments_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the child (student) to whom the reading will be assigned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentAssignReadingRequestDTO'
      responses:
        '201':
          description: Assignment created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponseDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Child or Reading not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Parent - Assignments
      summary: Parent List Child's Assignments
      description: Retrieves a list of all readings assigned by the parent to a specific child.
      operationId: parent_list_child_assignments_api_v1_parent_children__child_id__assignments_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the child (student) whose assignments are to be listed.
        # Add pagination parameters if applicable (page, size)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssignmentResponseDTO' # Or a more specific DTO for listing assignments
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Child not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parent/children/{child_id}/assignments/{assignment_id}:
    put:
      tags:
        - Parent - Assignments
      summary: Parent Update Child's Assignment
      description: Allows an authenticated parent to update details (e.g., due date) of an existing assignment for their child.
      operationId: parent_update_child_assignment_api_v1_parent_children__child_id__assignments__assignment_id__put
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the child (student).
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid # Assuming assignment_id is a UUID; adjust if it's an assessment_id
          description: The ID of the assignment to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentUpdateDTO'
      responses:
        '200':
          description: Assignment updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponseDTO' # Or updated assignment details
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Child or Assignment not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Parent - Assignments
      summary: Parent Delete Child's Assignment
      description: Allows an authenticated parent to delete a specific assignment for their child.
      operationId: parent_delete_child_assignment_api_v1_parent_children__child_id__assignments__assignment_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the child (student).
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid # Assuming assignment_id is a UUID
          description: The ID of the assignment to delete.
      responses:
        '204':
          description: Assignment deleted successfully. No content.
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Child or Assignment not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parent/children/{child_student_id}/progress:
    get:
      tags:
        - Parent - Child Monitoring
      summary: Parent Get Child Progress Summary
      description: Allows a parent to view the progress summary of one of their linked children.
      operationId: parent_get_child_progress_summary_api_v1_parent_children__child_student_id__progress_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the child (student) whose progress is to be viewed.
            title: Child Student Id
          description: The ID of the child (student) whose progress is to be viewed.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProgressSummaryDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parent/children/{child_student_id}/assessments/{assessment_id}/results:
    get:
      tags:
        - Parent - Child Monitoring
      summary: Parent Get Child Assessment Result Details
      description: Allows a parent to view detailed results of a specific assessment for one of their linked children.
      operationId: parent_get_child_assessment_result_details_api_v1_parent_children__child_student_id__assessments__assessment_id__results_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: child_student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the child (student).
            title: Child Student Id
          description: The ID of the child (student).
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the assessment.
            title: Assessment Id
          description: The ID of the assessment.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResultDetailDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssignmentResponseDTO: # This DTO might need to be more generic or have variants if used for parent assignments too
      properties:
        message:
          type: string
          title: Message
          default: Reading assigned successfully. # Consider if this default is always true
        created_assessments: # For parent assignment, this might be a single assessment
          items:
            $ref: '#/components/schemas/CreatedAssignmentInfoDTO'
          type: array # For parent, this might not be an array, or an array of 1
          title: Created Assessments
        skipped_students: # Likely not applicable for parent assigning to one child
          anyOf:
            - items:
                type: string
                format: uuid
              type: array
            - type: 'null'
          title: Skipped Students
          description: List of student IDs for whom assessment creation was skipped (e.g., invalid ID, not a student).
        # Consider adding fields specific to the assignment itself if needed, e.g., assignment_id, due_date
        # For parent context, it might be useful to return the specific assessment_id if only one is created.
      type: object
      title: AssignmentResponseDTO
      description: DTO for the response after a reading assignment. Can be by teacher or parent.
    AssignmentUpdateDTO:
      type: object
      properties:
        due_date:
          type: string
          format: date
          nullable: true
          description: New due date for the assignment.
        # Add other updatable fields if any, e.g., notes, status (if applicable)
      title: AssignmentUpdateDTO
      description: DTO for updating an existing assignment (e.g., by a parent for their child).
...
    CreatedAssignmentInfoDTO:
      properties:
        assessment_id:
          type: string
          format: uuid
          title: Assessment Id
        student_id:
          type: string
          format: uuid
          title: Student Id
        reading_id:
          type: string
          format: uuid
          title: Reading Id
        status:
          $ref: '#/components/schemas/AssessmentStatus'
      type: object
      required:
        - assessment_id
        - student_id
        - reading_id
        - status
      title: CreatedAssignmentInfoDTO
      description: DTO providing information about a single assessment created as part of an assignment.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ParentAssignReadingRequestDTO:
      type: object
      properties:
        reading_id:
          type: string
          format: uuid
          description: The ID of the reading material to assign.
        due_date:
          type: string
          format: date
          nullable: true
          description: Optional due date for the assignment.
      required:
        - reading_id
      title: ParentAssignReadingRequestDTO
      description: DTO for a parent to assign a reading to a specific child.
    ParentChildCreateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        preferred_language:
          type: string
          nullable: true
          default: 'en'
        role:
          type: string
          enum: [student] # Fixed to student
          default: student
      required:
        - email
        - password
      title: ParentChildCreateRequest
      description: Request body for a parent to create a new child (student) account.
...
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        first_name:
          anyOf:
            - type: string
            - type: 'null'
          title: First Name
        last_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Last Name
        preferred_language:
          anyOf:
            - type: string
            - type: 'null'
          title: Preferred Language
          default: en
        user_id:
          type: string
          format: uuid
          title: User Id
        role:
          type: string
          title: Role
      type: object
      required:
        - email
        - user_id
        - role
      title: UserResponse
    UserResponseDTO:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        email:
          type: string
          format: email
          title: Email
        first_name:
          anyOf:
            - type: string
            - type: 'null'
          title: First Name
        last_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Last Name
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
        - user_id
        - email
        - role
      title: UserResponseDTO
      description: |-
        A DTO for representing basic user details, often nested in other responses
        (e.g., details of students in a class list).
        It omits sensitive information like password_hash.
    UserRole:
      type: string
      enum:
        - student
        - teacher
        - parent
        - admin
      title: UserRole
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login